requirements.txt
streamlit
Pillow
opencv-python
numpy
